+++ src/MediaRecorder.js
@@ -1,87 +1,42 @@
-var MediaRecorder = function() {
-  this.initialize.call(this);
+navigator.getUserMedia = navigator.getUserMedia || navigator.webkitGetUserMedia || navigator.mozGetUserMedia;
+window.AudioContext = window.AudioContext || window.webkitAudioContext || window.mozAudioContext;
+window.URL = window.URL || window.webkitURL;
+
+// Feature detection.
+var getUserMediaCheck = typeof(navigator.getUserMedia) === 'function';
+var mediaRecorderCheck = typeof(window.MediaRecorder) === 'function';
+var webAudioCheck = typeof(window.AudioContext) === 'function';
+
+var Recorder = function(options) {
+  this.initialize.call(this, options);
 };
 
-MediaRecorder.prototype.initialize = function(cfg) {
-  window.AudioContext = window.AudioContext || window.webkitAudioContext;
-  navigator.getUserMedia = navigator.getUserMedia || navigator.webkitGetUserMedia;
-  if (!AudioContext) {
-    console.log('WebAudio not supported.');
+Recorder.getInstance = function(options) {
+  // Use the MediaRecorder API. Currently only works in firefox.
+  if (getUserMediaCheck && webAudioCheck && mediaRecorderCheck) {
+    Recorder = MediaRecorderAPI;
+  // Use HTML5 features (Web Audio API).
+  } else if (getUserMediaCheck && webAudioCheck && !mediaRecorderCheck) {
+    Recorder = MediaRecorder;
+    // Use Flash.
+  } else {
+    Recorder = MediaRecorderFlash;
   }
-  if (!navigator.getUserMedia) {
-    console.log('Navigator.getUserMedia not supported.');
-  }
-
-  this.audio_context = new AudioContext;
-  this.recLength = 0;
-  this.recBuffersL = [];
-  this.recBuffersR = [];
-  var config = cfg || {};
-  this.bufferLen = config.bufferLen || 4096;
-  this.sampleRate = this.audio_context.sampleRate;
-  navigator.getUserMedia({audio: true}, _.bind(this.prepare, this), function(e) {
-    console.log('No live audio input: ' + e);
-  });
+  return new Recorder(options);
 };
 
-MediaRecorder.prototype.prepare = function prepare(stream) {
-  this.input = this.audio_context.createMediaStreamSource(stream);
-  this.context = this.input.context;
-  this.node = (this.context.createScriptProcessor ||
-             this.context.createJavaScriptNode).call(this.context,
-                                                     this.bufferLen, 2, 2);
-  this.node.onaudioprocess = _.bind(this.onAudioProcess, this);
-  this.input.connect(this.node);
-  this.node.connect(this.context.destination);
+Recorder.prototype.initialize = function(cfg) {
 };
 
-MediaRecorder.prototype.onAudioProcess = function onAudioProcess(e) {
-  if (!this.recording) return;
-  var buffer = [new Float32Array(e.inputBuffer.getChannelData(0)), new Float32Array(e.inputBuffer.getChannelData(1))];
-  this.recBuffersL.push(buffer[0]);
-  this.recBuffersR.push(buffer[1]);
-  this.recLength += buffer[0].length;
+Recorder.prototype.record = function record() {
 };
 
-MediaRecorder.prototype.record = function record() {
-  this.clear();
-  this.recording = true;
+Recorder.prototype.clear = function clear() {
 };
 
-MediaRecorder.prototype.clear = function clear() {
-  this.recLength = 0;
-  this.recBuffersL = [];
-  this.recBuffersR = [];
+Recorder.prototype.play = function play() {
 };
 
-MediaRecorder.prototype.getBuffer = function getBuffer() {
-  var buffers = [];
-  buffers.push( this.mergeBuffers(this.recBuffersL, this.recLength) );
-  buffers.push( this.mergeBuffers(this.recBuffersR, this.recLength) );
-  return buffers;
-};
-
-MediaRecorder.prototype.mergeBuffers = function mergeBuffers(recBuffers, recLength){
-  var result = new Float32Array(recLength);
-  var offset = 0;
-  for (var i = 0; i < recBuffers.length; i++){
-    result.set(recBuffers[i], offset);
-    offset += recBuffers[i].length;
-  }
-  return result;
-};
-
-MediaRecorder.prototype.play = function play() {
-  var buffers = this.getBuffer();
-  var newBuffer = this.audio_context.createBuffer(2, buffers[0].length, this.audio_context.sampleRate);
-  this.outputSource = this.audio_context.createBufferSource();
-  newBuffer.getChannelData(0).set(buffers[0]);
-  newBuffer.getChannelData(1).set(buffers[1]);
-  this.outputSource.buffer = newBuffer;
-  this.outputSource.connect(this.audio_context.destination);
-  this.outputSource.start(0);
-};
-
-MediaRecorder.prototype.stop = function stop() {
+Recorder.prototype.stop = function stop() {
   this.recording = false;
 };
