+++ test/test.html
@@ -13,7 +13,10 @@
   <h2>Log</h2>
   <pre id="log"></pre>
 
-  <script src="../media-recorder.js"></script>
+  <script src="../src/MediaRecorder.js"></script>
+  <script src="../src/MediaRecorderFlash.js"></script>
+  <script src="../src/MediaRecorderAPI.js"></script>
+  <script src="../src/MediaRecorderHtml5.js"></script>
   <script>
   function __log(e, data) {
     log.innerHTML += "\n" + e + " " + (data || '');
@@ -35,7 +38,7 @@
   }
 
   window.onload = function init() {
-    recorder = new MediaRecorderFlash({swfSrc: '../recorder.swf'});
+    recorder = Recorder.getInstance({swfSrc: '../recorder.swf'});
     __log('Recorder initialised.');
   };
   </script>
